Release 1.9.7 & Fix bugs

+++ src/main/java/cn/mccraft/pangu/core/client/ui/builtin/MultiLineText.java
@@ -34,7 +34,7 @@ public class MultiLineText extends Component {
     @Setter
     protected FontProvider font;
 
-    public MultiLineText(int width, List<String> content, FontProvider font) {
+    public MultiLineText(float width, List<String> content, FontProvider font) {
         setSize(width, content.size() * (font.getFontHeight() + getPadding()));
         setFont(font);
         setContent(content);
